Refactor auth code
Nonfucntional change. Just reorganizing the code
