+++ pkg/auth/providers/providers.go
@@ -4,8 +4,11 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/rancher/norman/types"
 	"github.com/rancher/types/apis/management.cattle.io/v3"
 	"github.com/rancher/types/config"
+
+	"sync"
 
 	"github.com/rancher/rancher/pkg/auth/providers/github"
 	"github.com/rancher/rancher/pkg/auth/providers/local"
@@ -14,6 +17,7 @@ import (
 //Providers map
 var providers map[string]PrincipalProvider
 var providerOrderList []string
+var confMu sync.Mutex
 
 func init() {
 	providerOrderList = []string{"github", "local"}
@@ -25,6 +29,7 @@ type PrincipalProvider interface {
 	GetName() string
 	AuthenticateUser(input interface{}) (v3.Principal, []v3.Principal, map[string]string, int, error)
 	SearchPrincipals(name string, myToken v3.Token) ([]v3.Principal, int, error)
+	ConfigActionHandler(actionName string, action *types.Action, request *types.APIContext) error
 }
 
 func GetProvider(providerName string) (PrincipalProvider, error) {
@@ -37,6 +42,8 @@ func GetProvider(providerName string) (PrincipalProvider, error) {
 }
 
 func Configure(ctx context.Context, mgmtCtx *config.ManagementContext) {
+	confMu.Lock()
+	defer confMu.Unlock()
 	for _, providerName := range providerOrderList {
 		if _, exists := providers[providerName]; !exists {
 			switch providerName {
