+++ pkg/auth/providers/publicapi/handler.go
@@ -1,0 +1,61 @@
+package publicapi
+
+import (
+	"context"
+	"net/http"
+
+	normanapi "github.com/rancher/norman/api"
+	"github.com/rancher/norman/store/crd"
+	"github.com/rancher/norman/store/subtype"
+	"github.com/rancher/norman/types"
+	managementSchema "github.com/rancher/types/apis/management.cattle.io/v3/schema"
+	publicSchema "github.com/rancher/types/apis/management.cattle.io/v3public/schema"
+	v3public "github.com/rancher/types/client/management/v3public"
+	"github.com/rancher/types/config"
+)
+
+func NewHandler(ctx context.Context, mgmtCtx *config.ManagementContext) (http.Handler, error) {
+	schemas := types.NewSchemas().AddSchemas(publicSchema.PublicSchemas)
+	if err := authProviderSchemas(ctx, mgmtCtx, schemas); err != nil {
+		return nil, err
+	}
+
+	server := normanapi.NewAPIServer()
+	if err := server.AddSchemas(schemas); err != nil {
+		return nil, err
+	}
+	return server, nil
+}
+
+var authProviderTypes = []string{v3public.LocalProviderType, v3public.GithubProviderType}
+
+func authProviderSchemas(ctx context.Context, management *config.ManagementContext, schemas *types.Schemas) error {
+	schema := schemas.Schema(&publicSchema.PublicVersion, v3public.AuthProviderType)
+	setAuthProvidersStore(schema, management)
+	lh := &loginHandler{}
+	schema.ActionHandler = lh.login
+	schema.Formatter = loginActionFormatter
+
+	for _, apSubtype := range authProviderTypes {
+		subSchema := schemas.Schema(&publicSchema.PublicVersion, apSubtype)
+		subSchema.Store = subtype.NewSubTypeStore(apSubtype, schema.Store)
+		subSchema.ActionHandler = lh.login
+		subSchema.Formatter = loginActionFormatter
+	}
+
+	crdStore, err := crd.NewCRDStoreFromConfig(management.RESTConfig)
+	if err != nil {
+		return err
+	}
+
+	var crdSchemas []*types.Schema
+	for _, schema := range schemas.SchemasForVersion(managementSchema.Version) {
+		crdSchemas = append(crdSchemas, schema)
+	}
+
+	return crdStore.AddSchemas(ctx, crdSchemas...)
+}
+
+func loginActionFormatter(apiContext *types.APIContext, resource *types.RawResource) {
+	resource.AddAction(apiContext, "login")
+}
