+++ pkg/auth/tokens/apihandler.go
@@ -1,0 +1,61 @@
+package tokens
+
+import (
+	"context"
+	"net/http"
+
+	normanapi "github.com/rancher/norman/api"
+	"github.com/rancher/norman/store/crd"
+	"github.com/rancher/norman/types"
+	managementSchema "github.com/rancher/types/apis/management.cattle.io/v3/schema"
+	"github.com/rancher/types/client/management/v3"
+	"github.com/rancher/types/config"
+)
+
+var crdVersions = []*types.APIVersion{
+	&managementSchema.Version,
+}
+
+func NewAPIHandler(ctx context.Context, mgmtCtx *config.ManagementContext) (http.Handler, error) {
+	err := NewTokenAPIServer(ctx, mgmtCtx)
+	if err != nil {
+		return nil, err
+	}
+
+	schemas := types.NewSchemas().AddSchemas(managementSchema.TokenSchemas)
+
+	if err := tokenSchema(ctx, mgmtCtx, schemas); err != nil {
+		return nil, err
+	}
+
+	server := normanapi.NewAPIServer()
+
+	if err := server.AddSchemas(schemas); err != nil {
+		return nil, err
+	}
+
+	return server, nil
+}
+
+func tokenSchema(ctx context.Context, management *config.ManagementContext, schemas *types.Schemas) error {
+	schema := schemas.Schema(&managementSchema.Version, client.TokenType)
+	schema.CollectionActions = map[string]types.Action{
+		"logout": {},
+	}
+	schema.ActionHandler = tokenActionHandler
+	schema.ListHandler = tokenListHandler
+	schema.CreateHandler = tokenCreateHandler
+	schema.DeleteHandler = tokenDeleteHandler
+
+	crdStore, err := crd.NewCRDStoreFromConfig(management.RESTConfig)
+	if err != nil {
+		return err
+	}
+
+	var crdSchemas []*types.Schema
+	for _, schema := range schemas.SchemasForVersion(managementSchema.Version) {
+		crdSchemas = append(crdSchemas, schema)
+	}
+
+	return crdStore.AddSchemas(ctx, crdSchemas...)
+}
