+++ server/server.go
@@ -7,8 +7,9 @@ import (
 	"github.com/gorilla/mux"
 	k8sProxy "github.com/rancher/rancher/k8s/proxy"
 	managementapi "github.com/rancher/rancher/pkg/api/management/server"
-	"github.com/rancher/rancher/pkg/auth/filter"
-	"github.com/rancher/rancher/pkg/auth/server"
+	authfilter "github.com/rancher/rancher/pkg/auth/filter"
+	"github.com/rancher/rancher/pkg/auth/providers/publicapi"
+	"github.com/rancher/rancher/pkg/auth/tokens"
 	"github.com/rancher/rancher/pkg/remotedialer"
 	"github.com/rancher/rancher/pkg/tunnel"
 	"github.com/rancher/rancher/server/capabilities"
@@ -37,8 +38,8 @@ type Server struct {
 
 func New(ctx context.Context, httpPort, httpsPort int, management *config.ManagementContext) (*Server, error) {
 	var result http.Handler
-	tokenAPI, err := server.NewTokenAPIHandler(ctx, management)
-	publicAPI, err := server.NewAuthProviderAPIHandler(ctx, management)
+	tokenAPI, err := tokens.NewAPIHandler(ctx, management)
+	publicAPI, err := publicapi.NewHandler(ctx, management)
 
 	managementAPI, err := managementapi.New(ctx, httpPort, httpsPort, management, func() http.Handler {
 		return result
@@ -57,7 +58,7 @@ func New(ctx context.Context, httpPort, httpsPort int, management *config.Manage
 
 	authedAPIs := newAuthed(tokenAPI, managementAPI, k8sProxy)
 
-	authedHandler, err := filter.NewAuthenticationFilter(ctx, management, authedAPIs)
+	authedHandler, err := authfilter.NewAuthenticationFilter(ctx, management, authedAPIs)
 	if err != nil {
 		return nil, err
 	}
